requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
numpy==1.24.3
opencv-python-headless==4.7.0.72
Pillow==9.5.0
ultralytics==8.0.206
onnxruntime
onnx